重写下载 m3u8 逻辑

+++ app/src/main/java/euphoria/psycho/tasks/HLSDownloadTaskSegment.java
@@ -5,7 +5,7 @@ public class HLSDownloadTaskSegment {
     public String UniqueId;
     public String Uri;
     public int Sequence;
-    public int Total;
+    public long Total;
     public int Status;
     public long CreateAt;
     public long UpdateAt;
